refactor(imports): simplify deep relative imports
