+++ cms/src/api/foundation-navigation/content-types/foundation-navigation/lifecycles.ts
@@ -1,6 +1,5 @@
 import path from 'path'
-import { createNavigationLifecycle } from '../../../../utils/navigationLifecycle'
-import { PATHS } from '../../../../utils/paths'
+import { createNavigationLifecycle, PATHS } from '../../../utils'
 
 export default createNavigationLifecycle({
   contentTypeUid: 'api::foundation-navigation.foundation-navigation',
