+++ cms/src/api/summit-navigation/content-types/summit-navigation/lifecycles.ts
@@ -1,6 +1,5 @@
 import path from 'path'
-import { createNavigationLifecycle } from '../../../../utils/navigationLifecycle'
-import { PATHS } from '../../../../utils/paths'
+import { createNavigationLifecycle, PATHS } from '../../../utils'
 
 export default createNavigationLifecycle({
   contentTypeUid: 'api::summit-navigation.summit-navigation',
