+++ src/pages/developers/blog/[...id].astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection, render } from 'astro:content'
-import CommunityLinks from '../../../components/blog/CommunityLinks.astro'
-import BlogLayout from '../../../layouts/BlogLayout.astro'
+import CommunityLinks from '@/components/blog/CommunityLinks.astro'
+import BlogLayout from '@/layouts/BlogLayout.astro'
 
 export const prerender = true
