+++ src/pages/developers/blog/es.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from 'astro:content'
-import BaseLayout from '../../../layouts/BaseLayout.astro'
-import { createExcerpt } from '../../../utils/create-excerpt'
+import BaseLayout from '@/layouts/BaseLayout.astro'
+import { createExcerpt } from '@/utils/create-excerpt'
 
 // Only return posts with `draft: true` in the frontmatter
 const esEntries = await getCollection('engineering-blog', ({ data }) => {
