+++ src/pages/developers/blog/preview.astro
@@ -1,7 +1,7 @@
 ---
-import BlogLayout from '../../../layouts/BlogLayout.astro'
-import CommunityLinks from '../../../components/blog/CommunityLinks.astro'
-import { fetchStrapi } from '../../../utils/fetchStrapi'
+import CommunityLinks from '@/components/blog/CommunityLinks.astro'
+import BlogLayout from '@/layouts/BlogLayout.astro'
+import { fetchStrapi } from '@/utils/fetchStrapi'
 import { marked } from 'marked'
 
 // Get the documentId from the query string
